+++ frontend-plugin-core/src/main/java/com/github/eirslett/maven/plugins/frontend/lib/NodeExecutorConfig.java
@@ -1,6 +1,11 @@
 package com.github.eirslett.maven.plugins.frontend.lib;
 
+import java.io.BufferedReader;
 import java.io.File;
+import java.io.IOException;
+import java.io.InputStreamReader;
+import java.nio.charset.StandardCharsets;
+import java.util.Arrays;
 
 public interface NodeExecutorConfig {
   File getNodePath();
@@ -60,7 +65,108 @@ final class InstallNodeExecutorConfig implements NodeExecutorConfig {
   public File getInstallDirectory() {
     return installConfig.getInstallDirectory();
   }
-  
+
+  @Override
+  public File getWorkingDirectory() {
+    return installConfig.getWorkingDirectory();
+  }
+
+  @Override
+  public Platform getPlatform() {
+    return installConfig.getPlatform();
+  }
+}
+
+final class NodeSystemExecuterConfig implements NodeExecutorConfig {
+  private final String nodeExecutable;
+  private final String npmExecutable;
+
+
+  private final String npm;
+  private final String pnpm;
+  private final String pnpmCjs;
+  private final String npx;
+
+  private final InstallConfig installConfig;
+
+  public NodeSystemExecuterConfig(InstallConfig installConfig) {
+    this.installConfig = installConfig;
+    String nodeBinaryName = getPlatform().isWindows() ? "node.exe" : "node";
+    String environmentSeparator = getPlatform().isWindows() ? ";" : ":";
+    String npmCliName = getPlatform().isWindows() ? "npm.cmd" : "npm";
+
+    nodeExecutable = getPathOfExecutable(nodeBinaryName, environmentSeparator);
+    npmExecutable = getPathOfExecutable(npmCliName, environmentSeparator);
+
+
+    String systemLibDir = findGloballyInstalledPackages();
+
+
+    npm = systemLibDir + "/npm/bin/npm-cli.js";
+    pnpm = systemLibDir + "/pnpm/bin/pnpm.js";
+    pnpmCjs = systemLibDir + "/pnpm/bin/pnpm.cjs";
+    npx = systemLibDir + "/npm/bin/npx-cli.js";
+  }
+
+  private String findGloballyInstalledPackages() {
+    String systemLibDir;
+    try {
+      Runtime runtime = Runtime.getRuntime();
+      Process process = runtime.exec(new String[]{npmExecutable, "root", "-g"});
+      try (BufferedReader br = new BufferedReader(
+              new InputStreamReader(process.getInputStream(),
+              StandardCharsets.UTF_8))) {
+        systemLibDir = br.readLine().trim();
+      }
+      process.destroy();
+    } catch (IOException | NullPointerException e) {
+      throw new IllegalStateException("NPM cannot be executed!", e);
+    }
+    return systemLibDir;
+  }
+
+  private String getPathOfExecutable(String binaryName, String environmentSeparator) {
+    return Arrays.stream(System.getenv("PATH").split(environmentSeparator))
+            .map(File::new)
+            .filter(File::isDirectory)
+            .map(file -> new File(file.getAbsolutePath() + File.separator + binaryName))
+            .filter(File::exists)
+            .findFirst()
+            .orElseThrow(() -> new IllegalStateException(binaryName + " was not found on the system"))
+            .getAbsolutePath();
+  }
+
+  @Override
+  public File getNodePath() {
+    return new File(nodeExecutable);
+  }
+
+  @Override
+  public File getNpmPath() {
+    return new File(Utils.normalize(npm));
+  }
+
+
+  @Override
+  public File getPnpmPath() {
+    return new File(Utils.normalize(pnpm));
+  }
+
+  @Override
+  public File getPnpmCjsPath() {
+    return new File(Utils.normalize(pnpmCjs));
+  }
+
+  @Override
+  public File getNpxPath() {
+    return new File(Utils.normalize(npx));
+  }
+
+  @Override
+  public File getInstallDirectory() {
+    return new File(nodeExecutable);
+  }
+
   @Override
   public File getWorkingDirectory() {
     return installConfig.getWorkingDirectory();